+++ Pather.Common/Models/GameSegment/InitialGameUser.cs
@@ -6,8 +6,8 @@ namespace Pather.Common.Models.GameSegment
     public class InitialGameUser
     {
         public string UserId;
-        public int X;
-        public int Y;
+        public double X;
+        public double Y;
         public string GatewayId;
         public string GameSegmentId;
     }
